# Documentazione/OpenAPI.yaml
openapi: "3.0.2"
info:
  title: "Aulery"
  description: "Specifica progetto Aulery"
  version: "1.0"
  termsOfService: "https://aulery.it/terms"
  contact:
    name: "Aulery API by Ansima"
    url: "https://aulery.it/info"
    email: "contact@aulery.it"
  license:
    name: "CC Aulery1.0(CC AU 1.0)"
    url: "https://aulery.it"
...
servers:
  - url: "localhost/api/"
    description: Server LocalHost
components:
  schemas:
...
    Area:
      type: object
      properties:
        nome:
          type: string
        descrizione:
          type: string
    Corso:
      type: object
      properties:
        nome:
          type: string
        area:
          $ref: '#/components/schemas/Area'
    Evento:
      type: object
      properties:
        nome:
          type: string
        descrizione:
          type: string
        data:
          type: string
          format: date
        oraInizio:
          type: string
          format: time
        oraFine:
          type: string
          format: time
        tipologiaEvento:
          $ref: '#/components/schemas/TipologiaEvento'
        persona:
          $ref: '#/components/schemas/Persona'
        aula:
          $ref: '#/components/schemas/Aula'
        corso:
          $ref: '#/components/schemas/Corso'
        eventoRicorrente:
          $ref: '#/components/schemas/EventoRicorrente'
    EventoRicorrente:
      type: object
      properties:
        dataFine:
          type: string
          format: date
        ricorrenza:
          $ref: '#/components/schemas/Ricorrenza'
    Aula:
      type: object
      properties:
        nome:
          type: string
        note:
          type: string
        edificio:
          $ref: '#/components/schemas/Edificio'
        piano:
          type: integer
        zona:
          type: string
        capienza:
          type: integer
        preseElettriche:
          type: integer
        preseDiRete:
          type: integer
        attrezzature:
          type: array
          items:
            $ref: '#/components/schemas/Attrezzatura'
        area:
          $ref: '#/components/schemas/Area'
    Persona:
      type: object
      properties:
        nome:
          type: string
        cognome:
          type: string
        email:
          type: string  
    Attrezzatura:
      type: string
      enum:
        - PROIETTORE
        - SCHERMOMOTORIZZATO
        - SCHERMOMANUALE
        - IMPIANTOAUDIO
        - PCFISSO
        - MICROFONOAFILO
        - MICROFONOSENZAFILO
        - LAVAGNALUMINOSA
        - WIFI
    Edificio:
      type: string
      enum:
        - COPPITO0
        - COPPITO1
        - COPPITO2
        - ALANTURING
        - BLOCCOA
        - BLOCCOB
        - BLOCCOC
    Ricorrenza:
      type: string
      enum:
        - GIORNALIERA
        - SETTIMANALE
        - MENSILE
        - ANNUALE  
    TipologiaEvento:
      type: string
      enum:
        - LEZIONE
        - ESERCITAZIONE
        - ESAME
        - PARZIALE
        - LAUREE
        - SEMINARIO
        - CONFERENZA
        - RIUNIONE
        - CDC
        - CAD
        - CERIMONIA
        - ELEZIONI
        - DIPARTIMENTO
        - ATENEO
        - ESTERNO
        - MANUTENZIONE
        - STUDIOINDIVIDUALE
        - ALTRO
  parameters:
...
    aulaID:
      name: aulaID
      in: path
      description: Id dell'aula
      required: true
      schema:
        type: integer
    areaID:
      name: areaID
      in: path
      description: Id dell'area
      required: true
      schema:
        type: integer
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: internalToken
  responses:
    GeneralError:
      description: Errore 500, Errore interno
      content:
        text/plain:
          schema:
            type: string
            example: "Errore 500, Errore interno"
...
    NotFoundError:
      description: Errore 404, Elemento non trovato
      content:
        text/plain:
          schema:
            type: string
            example: "Errore 404, Elemento non trovato"   
paths:
...
  /rest/aule/esportazione:
    get:
      summary: Esportazione aule
      description: Operazione che permette di esportare un file CSV contenente una lista di tutte le aule.
      operationId: esportaAuleCSV
      tags:
        - Aule
      security:
        - bearerAuth: []
      responses:
        200:
          description: Esportazione riuscita con successo
          content:
            text/csv:
              schema:
                type: string
                format: binary
        404:
          $ref: '#/components/responses/NotFoundError'
        500:
          $ref: '#/components/responses/GeneralError'
  /rest/aule/importazione:
    post:
      summary: Importazione aule
      description: Operazione che permette di importare un file CSV contenente una lista di aule.
      operationId: importaAuleCSV
      tags:
        - Aule
      security:
        - bearerAuth: []
      requestBody:
        content:
          text/csv:
            schema:
              type: string
              format: binary
      responses:
        200:
          description: Importazione riuscita
        404:
          $ref: '#/components/responses/NotFoundError'
        500:
          $ref: '#/components/responses/GeneralError'
  /rest/aule/crea:
    post:
      summary: Crea aula
      description: Operazione che permette di aggiungere una aula al sistema.
      tags: 
        - Aule
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Aula'
      responses:
        201:
          description: Aula creata con successo
          headers:
            location:
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                type: string
                format: uri
        404:
          $ref: '#/components/responses/NotFoundError'
        500:
          $ref: '#/components/responses/GeneralError'
  /rest/aule/{aulaID}/assegnaArea:
    put:
      summary: Assegna area ad aula
      description: Operazione che permette di assegnare una specifica area ad una aula già presente nel sistema.
      operationId: assegnaArea
      tags:
        - Aule
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/aulaID'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                areaId:
                  type: integer
                  description: ID dell'area
      responses:
        200:
          description: Assegnazione riuscita
        404:
          $ref: '#/components/responses/NotFoundError'
        500:
          $ref: '#/components/responses/GeneralError'
  /rest/aule/{aulaID}/attrezzature:
    get:
      summary: Lista attrezzature aula
      description: Operazione che permette di ottenere una lista di attrezzature di una aula già presente nel sistema.
      operationId: getAttrezzature
      tags: 
        - Aule
      parameters:
      - $ref: '#/components/parameters/aulaID'
      responses:
        200:
          description: Lista ottenuta con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Aula'
        404:
          $ref: '#/components/responses/NotFoundError'
        500:
          $ref: '#/components/responses/GeneralError'
  /rest/aula/{aulaID}/info:
    get:
      summary: Info aula
      description: Operazione che permette di ottenere le informazioni riguardo una aula già presente nel sistema.
      operationId: getAula
      tags:
        - Aule
      parameters:
      - $ref: '#/components/parameters/aulaID'
      responses:
        '200':
          description: Informazioni ottenute con successo 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Aula'
        404:
          $ref: '#/components/responses/NotFoundError'
        500:
          $ref: '#/components/responses/GeneralError'
...
  /rest/collezioni/{areaID}/prossime3ore:
    get:
      summary: Lista eventi delle prossime tre ore
      description: Operazione che permette di ottenere una lista di eventi attuali e delle prossime tre ore di una intera area.
      operationId: getEventiTreOre
      tags: 
        - Liste
      parameters:
      - $ref: '#/components/parameters/areaID'
      responses:
        200:
          description: Lista ottenuta con successo
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Evento'
        404:
          $ref: '#/components/responses/NotFoundError'  
        500:
          $ref: '#/components/responses/GeneralError'